project/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
importlib-metadata==3.7.0
ipstack==0.1.4
Markdown==3.3.4
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
zipp==3.4.0
psycopg2-binary==2.8.5
gunicorn==20.0.4